documentation updates, queryable docs
